Added generic dependency tracking
The database objects have now a depend_on property. After the objects are
read from the schema, depend_on is populated by a list of (cls, key) pairs,
where cls is the Python class representing that object and key the instance
key.

The objects dicts now also index each object by oid, not only by key; the
map is in the by_oid dict.

Currently not doing anything with this map, and probably it contains
excessive information. The idea is to replace the ad-hoc dependency tracker
every class has grown with this generic map. The first step however will be
to emit the dependency map information in the yaml to analyze it better.