Sketching the framework for a topological sort
Dropped the a-priori (and broken) knowledge of the "right order" of the
classes to restore. The `diff_map()` method is split in an internal
version returning a map object -> changes so that we will be (hopefully)
able to topo-sort the statements according to the objects dependencies.
The refactoring is not implemented for all the objects: for the moment
I'm testing with a domain -> constraint -> function dependency as per
issue #100. It's also nice that the public schema ends up last in the
generated statements: the sort will have to put it first, even if it
doesn't make any diff).

The DbObject type is now hashable so that we can use it as dict key.