Pass the db around explicitly when getting the dependencies
This allows to implement a single function where each object
can return the database object it knows it depends on. The same
function is used both when getting the dependencies to walk the
objects graph and when storing the yaml to avoid saving redundant
information.

A `db` parameter has been added to `_base_map()` and all the
`to_map()` methods.