Don't attemp to dump temporary data used in dependency tracking
Code smell: every time we attach a random sticker on the object Pyrseas
tries to dump it to yaml. I think the approach "dump this attribute
unless we tell you not" is wrong and it should be instead "only dump
these attribute".

At least I think we may avoid dumping attributes starting with '_'.