Fixed dependency between tables and sequences
Adding the dependency of the defaults creates loops because the sequence
used to depend on the owner table but now the table defaults often
depends on the sequence. Toposort is not amused. So don't imply the
dependency of the seq on the owning table: make sure that the sequence
is always created before the table and have the OWNED BY statement
emitted by the table instead of the sequence, so it will definitely
happen after both have been created.