Dropped dependency set by the table on the constrants
They were only used to get the table name. We can do without.
This was part of an attemp to kill the entire link_refs section but it's
actually a die hard, so I'll take a different approach: see who gets in
the way and kill attribute by attribute until there is no tangle that
stops the global dependency tracking. These functions that create my
table, for instance...