Conditionally inject dependencies of functions
For some functions, it is possible that the functions it depends
upon are not explicitly present, e.g. for the following aggregate
function:

CREATE AGGREGATE array_agg_mult(anyarray) (
    SFUNC = array_cat,
    STYPE = anyarray,
    INITCOND = '{}'
);

We got the following error when trying to dbtoyaml this database

KeyError: (u'public', 'array_cat', u'anyarray, anyarray'), but
the dependency is upon pg_catalog.array_cat, so should not even be
made explicit.