Quote more keywords in SQL generation
Change filter to use any keywords not marked as non-reserved instead of
only those marked as reserved. This is necessary due to how certain
keywords are classified by Postgres and by the various standards.

Keyword categories (from Postgres `src/include/common/keywords.h`):
- `U` / `UNRESERVED_KEYWORD`
- `C` / `COL_NAME_KEYWORD`
- `T` / `TYPE_FUNC_NAME_KEYWORD`
- `R` / `RESERVED_KEYWORD`

For example, "authorization" is reserved, but may be used as a function
name, so it has a `catcode` of `T` instead of `R`, so is not quoted by
the original logic.

Docs: https://www.postgresql.org/docs/current/sql-keywords-appendix.html

Closes #212